move default_storageclasses fixture, fix imports
